Add time measurement

+++ roboteam_ai/src/utilities/Constants.h
@@ -25,7 +25,7 @@ const double ROBOT_RADIUS=0.089; // TODO: Need to test if world_state agrees wit
 const double DRIBBLER_ANGLE_OFFSET=asin(FRONT_LENGTH/2/ROBOT_RADIUS);
 const double MAX_BALL_RANGE=0.03; // Could maybe be even less? TODO: needs to be tested.
 const double DRIBBLE_POSDIF=0.03;
-const double DRIBBLE_SPEED=1.0;
+const float  DRIBBLE_SPEED=1.0;
 
 //Other/multiple usage
 const int DEFAULT_ROBOT_ID = 1;
